Start moving common nunchuk data into core::nunchuk
