+++ wii-ext/src/core.rs
@@ -1,1 +1,2 @@
 pub mod classic;
+pub mod nunchuk;
